+++ raiden-dapp/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [#1212] Allows user to view history of all transactions that has been made.
+- [#1380] Allow the user to withdraw tokens from the Raiden Account.
 - [#1365] Allow user to transfer ETH between Main and Raiden account.
 - [#1424] Sort tokens in token list by balance and symbol.
 
@@ -22,6 +23,7 @@
 
 - [#1458] General screen has been renamed to account.
 
+[#1380]: https://github.com/raiden-network/light-client/issues/1380
 [#1458]: https://github.com/raiden-network/light-client/issues/1458
 [#1402]: https://github.com/raiden-network/light-client/issues/1402
 [#1413]: https://github.com/raiden-network/light-client/issues/1413